+++ src/interafces/extensions/IExtensionRepositoryDescription.php
@@ -1,0 +1,13 @@
+<?php
+namespace extas\interfaces\extensions;
+
+/**
+ * Interface IExtensionRepositoryDescription
+ *
+ * @package extas\interfaces\extensions
+ * @author jeyroik@gmail.com
+ */
+interface IExtensionRepositoryDescription
+{
+
+}